Add config option to adjust the size of the holo display
Config is called renderScaling. The description says 0.0-1.0, but really
the minimum value is 0.1 (because 0 and you can barely see it at even
0.2). The maximum value is also unbounded by the code.

I initially tried to adjust the blockScaleModifier, but that didn't take
(the factor is used in lots of places so maybe I missed something), so I
pushed the scale matrix earlier in the stack instead. Maybe heavy-handed
but it does the job, and it adjusts text scaling too.